Tweak the "cog as library" API to accept language-specific options and return codejen.Files instead of bytes
